+++ client.d.ts
@@ -1,0 +1,4 @@
+﻿declare module 'virtual:color-suite-colors' {
+  const colors: {[key: string]: string | {[key: string]: string }}
+  export default colors
+}